+++ galacteek/application/__init__.py
@@ -92,6 +92,7 @@ from galacteek.browser.schemes import MultiObjectHostSchemeHandler
 from galacteek.browser.schemes.ipid import IPIDSchemeHandler
 from galacteek.browser.schemes.i import ISchemeHandler
 from galacteek.browser.schemes.gemini import GeminiSchemeHandler
+from galacteek.browser.schemes.gemini import GemIpfsSchemeHandler
 from galacteek.browser.schemes.pronto import ProntoGraphsSchemeHandler
 
 from galacteek.browser import BrowserRuntimeObjects
@@ -1594,6 +1595,7 @@ class GalacteekApplication(QApplication):
         self.ipidSchemeHandler = IPIDSchemeHandler(self)
         self.iSchemeHandler = ISchemeHandler(self)
         self.geminiSchemeHandler = GeminiSchemeHandler(self)
+        self.gemIpfsSchemeHandler = GemIpfsSchemeHandler(self)
         self.prontoGSchemeHandler = ProntoGraphsSchemeHandler(self)
 
     def subUrl(self, path):