+++ galacteek/browser/schemes/gemini/__init__.py
@@ -1,9 +1,11 @@
+import re
 import ignition
 from yarl import URL
 
 from PyQt5.QtCore import QUrl
 
 from galacteek import log
+from galacteek.ipfs import ipfsOp
 from galacteek.browser.schemes import BaseURLSchemeHandler
 from galacteek.browser.schemes import SCHEME_GEMINI
 
@@ -14,7 +16,19 @@ class GeminiError(Exception):
     pass
 
 
-class GeminiSchemeHandler(BaseURLSchemeHandler):
+class GeminiClient:
+    def geminiRequest(self, url: str):
+        # Run in the thread executor
+        try:
+            response = ignition.request(url)
+            data = response.data()
+            return response, data
+        except Exception as err:
+            log.debug(f'Gemini request error for URL {url}: {err}')
+            return None, None
+
+
+class GeminiSchemeHandler(BaseURLSchemeHandler, GeminiClient):
     """
     Simple Gemini URL scheme handler.
 
@@ -28,16 +42,6 @@ class GeminiSchemeHandler(BaseURLSchemeHandler):
         ignition.set_default_hosts_file(
             str(self.app.geminiHostsLocation)
         )
-
-    def geminiRequest(self, url: str):
-        # Run in the thread executor
-        try:
-            response = ignition.request(url)
-            data = response.data()
-            return response, data
-        except Exception as err:
-            log.debug(f'Gemini request error for URL {url}: {err}')
-            return None, None
 
     async def handleRequest(self, request, uid):
         rUrl = request.requestUrl()
@@ -152,3 +156,101 @@ class GeminiSchemeHandler(BaseURLSchemeHandler):
             log.debug(f'{rMethod}: {url}: error rendering capsule: {err}')
 
             return self.reqFailed(request)
+
+
+class GemIpfsSchemeHandler(BaseURLSchemeHandler, GeminiClient):
+    """
+    Gemini IPFS gateway scheme handler
+    """
+
+    def __init__(self, parent=None, noMutexes=False):
+        super().__init__(parent=parent, noMutexes=noMutexes)
+
+        # Set the default gemini known hosts file location
+        ignition.set_default_hosts_file(
+            str(self.app.geminiHostsLocation)
+        )
+
+    @ipfsOp
+    async def handleRequest(self, ipfsop, request, uid):
+        rUrl = request.requestUrl()
+        rMethod = bytes(request.requestMethod()).decode()
+
+        try:
+            parts = rUrl.path().lstrip('/').split('/')
+            host = parts[0]
+            capsule = parts[1]
+            rest = '/'.join(parts[2:])
+            path = rest if rest else '/'
+
+            assert re.match(r'[a-zA-Z0-9]+', host) is not None
+            assert re.match(r'[a-zA-Z0-9]+', capsule) is not None
+        except Exception:
+            return self.urlInvalid(request)
+
+        if not host or not capsule:
+            return self.urlInvalid(request)
+
+        p2pEndpoint = f'/p2p/{host}/x/gemini/{capsule}/1.0'
+
+        # Tunnel
+        async with ipfsop.p2pDialerFromAddr(p2pEndpoint,
+                                            allowLoopback=True) as dial:
+            if dial.failed:
+                return self.reqFailed(request)
+
+            if rUrl.hasQuery():
+                q = rUrl.query(QUrl.EncodeSpaces)
+                url = ignition.url(
+                    f'{path}?{q}',
+                    f'//{dial.maddrHost}:{dial.maddrPort}'
+                )
+            else:
+                url = ignition.url(
+                    path,
+                    f'//{dial.maddrHost}:{dial.maddrPort}'
+                )
+
+            # Run the request in the app's executor
+            response, data = await self.app.loop.run_in_executor(
+                self.app.executor,
+                self.geminiRequest,
+                url
+            )
+
+            if not response or not data:
+                return self.reqFailed(request)
+
+            meta = response.meta
+
+            if isinstance(data, bytes) and meta:
+                # Raw file
+
+                return self.serveContent(
+                    request.reqUid,
+                    request,
+                    meta,
+                    data
+                )
+
+            try:
+                if not response.success():
+                    raise GeminiError(
+                        f'{response.url}: Invalid response: {response.status}')
+
+                html, title = gemTextToHtml(data)
+
+                if not html:
+                    raise GeminiError(f'{response.url}: gem2html failed')
+
+                await self.serveTemplate(
+                    request,
+                    'gemini_capsule_render.html',
+                    gembody=html,
+                    gemurl=rUrl.toString(),
+                    title=title if title else rUrl.toString()
+                )
+            except Exception as err:
+                log.debug(f'{rMethod}: {url}: error rendering capsule: {err}')
+
+                return self.reqFailed(request)