+++ galacteek/docs/manual/__init__.py
@@ -1,1 +1,1 @@
-__manual_en_version__ = '20210310'
+__manual_en_version__ = '20210919'