+++ share/translations/galacteek_fr.ts
@@ -1621,7 +1621,7 @@ Path: {0}, nodes processed: {1}</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="135"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="142"/>
         <source>
         &lt;p&gt;
         Create a simple website on the dweb.
@@ -1633,98 +1633,103 @@ Path: {0}, nodes processed: {1}</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="149"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="156"/>
         <source>Open latest content in the pyramid</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="156"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="163"/>
         <source>Pyramid is empty</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="162"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="169"/>
         <source>Pop an item off the pyramid</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="168"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="175"/>
         <source>Force sync</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="174"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="181"/>
         <source>Rewind DAG</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="180"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="187"/>
         <source>Publish</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="187"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="194"/>
         <source>Unpublish</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="194"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="201"/>
         <source>Register this content as a service in the list of IP services on your DID (Decentralized Identity)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="201"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="208"/>
         <source>Rewinding the DAG cancels the latest operation/transformation in the DAG&apos;s history</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="208"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="215"/>
         <source>Copy IPNS address to clipboard</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="214"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="221"/>
         <source>Copy IPNS gateway address to clipboard</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="220"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="227"/>
         <source>Add current clipboard item to the pyramid</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="227"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="234"/>
         <source>Generate pyramid&apos;s QR code</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="234"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="241"/>
         <source>Generate index&apos;s QR code</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="241"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="248"/>
         <source>Choose a pyramid to publish this object to</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="248"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="255"/>
         <source>Drop object to a pyramid</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="255"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="262"/>
         <source>Browse image gallery</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="262"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="269"/>
         <source>Browse image gallery (IPNS)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="269"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="276"/>
         <source>Change image gallery&apos;s title</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../galacteek/ui/pyramids.py" line="135"/>
+        <source>Create: gem (gemini ipfs capsule)</source>
         <translation type="unfinished"></translation>
     </message>
 </context>